API_plans/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
pydantic
requests   # optional, for seed or local testing
